KTL-4710: move the search-eval harness into a dedicated e2e module and pin the regression floor to its snapshot key
