+++ e2e/module.yaml
@@ -1,14 +1,16 @@
-product: jvm/app
+product: jvm/lib
 
 apply:
   - ../build-logic/templates/spring-web.module-template.yaml
   - ../build-logic/templates/persistence.module-template.yaml
   - ../build-logic/templates/mock.module-template.yaml
-  - ../build-logic/templates/spring-scheduling.module-template.yaml
   - ../build-logic/templates/spring-cloud.module-template.yaml
-  - ../build-logic/templates/maven-indexer.module-template.yaml
+  - ../build-logic/templates/testcontainers.module-template.yaml
 
+# Mirrors app's dependencies: the tests boot the whole Application, whose component
+# scan over `io.klibs` instantiates beans from every module app depends on.
 dependencies:
+  - ../app
   - ../core/package
   - ../core/project
   - ../core/readme
@@ -17,22 +19,9 @@ dependencies:
   - ../core/search
   - ../core/storage
   - ../integrations/ai
-  - ../integrations/maven
   - ../integrations/github
+  - ../integrations/maven
   - ../integrations/mcp
 
 test-dependencies:
-  - $libs.okhttp
-  - $libs.kohsuke.githubApi
-
-settings:
-  jvm:
-    mainClass: io.klibs.app.ApplicationKt
-
-plugins:
-  jib:
-    enabled: true
-    baseImage:
-      fullName:  eclipse-temurin:21-jre
-    container:
-      mainClass: io.klibs.app.ApplicationKt
+  - $libs.spring.boot.testcontainers
