+++ e2e/src/test/resources/application-test.yml
@@ -1,0 +1,24 @@
+klibs:
+  indexing: false
+  ai: false
+  integration:
+    github:
+      personal-access-token: test
+  scheduling:
+    process-indexing-queue:
+      enabled: false
+    fill-version-type:
+      enabled: false
+
+spring:
+  sql:
+    init:
+      mode: always
+  cloud:
+    aws:
+      region:
+        static: eu-west-1
+      credentials:
+        access-key: test
+        secret-key: test
+      endpoint: http://localhost:4566
