+++ e2e/src/test/resources/search-eval/floor.json
@@ -1,0 +1,4 @@
+{
+  "ids" : [ "A1", "A10", "A11", "A12", "A13", "A14", "A2", "A3", "A4", "A5", "A6", "A7", "A8", "A9", "B1", "B10", "B12", "B13", "B14", "B16", "B17", "B18", "B20", "B21", "B25", "B26", "B27", "B29", "B3", "B30", "B31", "B32", "B33", "B34", "B35", "B36", "B37", "B38", "B39", "B4", "B40", "B41", "B43", "B44", "B45", "B46", "B47", "B48", "B49", "B50", "B51", "B52", "B53", "B54", "B55", "B56", "B57", "B58", "B59", "B60", "B61", "B8", "B9", "C10", "C2", "C7", "C8", "C9", "D1", "D2", "D3", "D4", "D5", "E2", "E3", "E4", "E6", "M1", "M11", "M12", "M3", "M4", "M5", "M6", "M7", "M8" ],
+  "snapshot" : "search-eval/frozen-2026-07-18.pgdump.gz"
+}
